feat(runtime): extract /auth and /ai dispatcher domain bodies — ADR-0076 D11 step ③ PR-7 (#2462)
- domains/auth.ts: better-auth service bridge + the MSW/browser mock
  fallback; the local browser-safe randomUUID wrapper moves with its
  only consumer.
- domains/ai.ts: dispatches the AI plugin's kernel-cached __aiRoutes
  table, enforcing each route's declared auth contract and threading
  the resolved actor. Receives the FULL cleanPath (legacy branch passed
  cleanPath whole; the matcher re-prefixes /api/v1) — preserved.
- DomainHandlerDeps grows isAuthRequired() (lazily read — the
  requireAuth field initializes AFTER the deps object, a construction-
  order trap TS2729 caught) and getRegisteredAiRoutes().
- /mcp deliberately EXCLUDED: buildMcpBridge couples to the
  action-execution family (callData/actionPermissionError/
  invokeBusinessAction/…), so the whole MCP family goes with the
  /actions /meta /data deep-coupling batch.

Verified: seam suite 42 tests, runtime 647 green, http-conformance 41
green, dependent closure builds with DTS (--force).
